Properly handle motion-notify-event (from lodatom).
When using POINTER_MOTION_HINT_MASK, only the first motion-notify-event
is fired, and then following events are suppressed until a button is
pressed or gtk.gtk.Window.get_pointer() is called.  The previous version
did not call get_pointer(), so on some systems, the event only fired
once, and the effect was that panning and the mouse-over effect did not
work.

Fixes issue #24.